requirements.txt
numpy==1.19.2
pandas==1.1.5
pytorch==1.5.1
scipy==1.5.2
scikit-learn==0.24.2
cudatoolkit==10.2.89
torchvision==0.6.0
tensorboard==2.4.0
torchsampler
